requirements.txt
numpy
torch
gym
gym[box2d]
pygame_gui
pygame
screeninfo
neat-python